[featue] routes namespace, controller調整
