+++ app/views/admin/users/edit.html.erb
@@ -1,5 +1,5 @@
 <div class="flex items-center">
-  <%= link_to '/admin' do %>
+  <%= link_to admin_path do %>
       <%= image_tag "/images/right-arrow.png", class: "w-4 h-4" %>
   <% end %>
   <h1 class="ml-4">
