+++ app/views/admin/users/index.html.erb
@@ -1,5 +1,5 @@
 <div class="flex content-center justify-between">
-  <%= link_to t('new_user'), admin_new_user_path, class: "text-white bg-red-400 p-2 px-8 rounded-full" %>
+  <%= link_to t('new_user'), new_admin_user_path, class: "text-white bg-red-400 p-2 px-8 rounded-full" %>
 </div>
 <% if flash.any? %>
     <% flash.each do |type, msg| %>
@@ -21,7 +21,7 @@
       <td><%= user.role %></td>
       <td><%= user.tasks.count %></td>
       <td>
-        <%= link_to admin_edit_user_path(user) do %>
+        <%= link_to edit_admin_user_path(user) do %>
           <%= image_tag "/images/edit.png", class: "w-4 h-4" %>
         <% end %>
       </td>
@@ -31,7 +31,7 @@
         <% end %>
       </td>
       <td>
-      <%= button_to image_tag("/images/close.png", class: "w-4 h-4"), admin_path(id: user.id), method: :delete,form: {data: {turbo_confirm: 'Are you sure?'}} %>
+      <%= button_to image_tag("/images/close.png", class: "w-4 h-4"), admin_user_path(id: user.id), method: :delete,form: {data: {turbo_confirm: 'Are you sure?'}} %>
       </td>
     </tr>
     <% end %>
